Moved trace plotting scripts into individual files
